Add UI frame builder for HUD snapshots
