+++ README.md
@@ -20,6 +20,7 @@ This repository has been reset to develop a real-time Apollo 11 mission simulato
 - Data validation CLI in [`js/src/tools/validateMissionData.js`](js/src/tools/validateMissionData.js) (`npm run validate:data`) sweeps the mission datasets for GET formatting issues, prerequisite/reference integrity, autopilot asset health, PAD parameter JSON validity, consumables budget sanity checks, nested communications metrics, and EVA extension flags.
 - Milestone M2 planning notes outline guidance, RCS, and docking system requirements in [`docs/milestones/M2_GUIDANCE_RCS.md`](docs/milestones/M2_GUIDANCE_RCS.md) to steer upcoming implementation work.
 - Text HUD snapshots generated by [`js/src/hud/textHud.js`](js/src/hud/textHud.js) now surface mission progress, resource margins, propellant usage, T-minus countdowns for upcoming events, active checklist chips, autopilot focus, and caution/warning counts at configurable intervals during CLI runs, aiding long-session monitoring.
+- `UiFrameBuilder` in [`js/src/hud/uiFrameBuilder.js`](js/src/hud/uiFrameBuilder.js) consolidates scheduler, resource, autopilot, checklist, and manual queue state into deterministic `ui_frame` snapshots (documented in [`docs/ui/ui_frame_reference.md`](docs/ui/ui_frame_reference.md)) so the CLI HUD and upcoming Navigation/Controls/Systems views share a single data contract.
 - Milestone M3 UI, HUD, and audio telemetry planning in [`docs/milestones/M3_UI_AUDIO.md`](docs/milestones/M3_UI_AUDIO.md) defines presentation-layer architecture, cue taxonomy, and accessibility handoff targets for the JS prototype and N64 port.
 - Detailed HUD and pane layout specification in [`docs/ui/hud_layout.md`](docs/ui/hud_layout.md) locks in the Always-On HUD modules, Navigation/Controls/Systems view grids, tile-mode presets, and data-binding strategy that the M3 UI build will implement across the web and N64 targets.
 - Milestone M4 N64 port plan in [`docs/milestones/M4_N64_PORT.md`](docs/milestones/M4_N64_PORT.md) maps the libdragon architecture, rendering/audio budgets, input scheme, and asset pipeline for the hardware build.
